+++ Food_Management_System.Domain/Entities/Inventory.cs
@@ -6,7 +6,7 @@ namespace Food_Management_System.Domain.Entities
     {
         public string ItemName { get; set; }
         public string Unit { get; set; }
-        public int QuantityAvailable { get; set; }
+        public double QuantityAvailable { get; set; }
         public int ReorderLevel { get; set; }
         public ICollection<Recipe>? Recipes { get; set; }
     }
